refactor: refactoring products filter to mount filters data separately
