+++ src/app/components/ProductsFilters/ProductsFilter.jsx
@@ -2,31 +2,15 @@ import { useEffect, useState } from "react";
 import { Button, Checkbox, FormControlLabel, Grid, Stack } from "@mui/material";
 import { Accordion, AccordionSummary, AccordionDetails } from "@mui/material";
 import AddIcon from '@mui/icons-material/Add';
-import { formatCurrency } from "../util"
+import { formatCurrency } from "../../util"
+import { setupFiltersData } from "./setupFiltersData";
 
-// TODO: refactor, separate the setup of filter data from component mounting
 export default function ProductsFilter(props) {
     const { products, handleFilter } = props;
 
     const [productsFilters, setProductsFilters] = useState([]);
     const [activeFilters, setActiveFilters] = useState({});
 
-    function getFilterFieldValues(fields) {
-        const fieldsWithValues = fields.map(field => {
-            let values = products.map(product => {
-                return product[field]
-            })
-            values = removeDuplicatedValues(values)
-            values = removeFalsyValues(values)
-
-            return { field, values }
-        })
-        return fieldsWithValues
-    }
-
-    function removeFalsyValues(values) {
-        return values.filter(value => !!value)
-    }
 
     function getTranslatedField(field) {
         const translatedToPortugueseFields = {
@@ -40,38 +24,17 @@ export default function ProductsFilter(props) {
         return translatedToPortugueseFields[field]
     }
 
-    function excludeNotFilterableFields(fields) {
-        const excludedFields = [
-            "createdAt",
-            "updatedAt",
-            "quantity",
-            "description",
-            "id",
-            "image",
-        ]
-        return fields.filter(field => !excludedFields.includes(field))
-    }
-
-    function removeDuplicatedValues(values) {
-        return [...new Set([...values])]
-    }
-
-    function getProductsFilters() {
-        let fields = products.reduce((accProducts, product) => {
-            return [...accProducts, ...Object.keys(product)]
-        }, [])
-
-        const uniqueFields = removeDuplicatedValues([...fields])
-        const filterableFields = excludeNotFilterableFields(uniqueFields)
-        const fieldsWithFilterableValues = getFilterFieldValues(filterableFields)
-        setProductsFilters(fieldsWithFilterableValues)
-    }
-
     useEffect(() => {
         if (products.length)
-            getProductsFilters()
+            setProductsFilters(setupFiltersData(products))
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [products])
+
+    useEffect(() => {
+        handleFilter(activeFilters)
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [activeFilters])
+
 
     function formatLabelByField(field, value) {
         if (!["discount", "price"].includes(field)) {
@@ -80,26 +43,28 @@ export default function ProductsFilter(props) {
         return field === "discount" ? `${value}%` : formatCurrency(value);
     }
 
-    // TODO: divide in separeted in functions
     function handleFieldChange(event) {
         const [field, value] = event.target.value.split("-")
         setActiveFilters(oldActiveFilters => {
             const fieldValues = oldActiveFilters[field] ? Object.values(oldActiveFilters[field]) : [];
             if (event.target.checked) {
-                oldActiveFilters[field] = [...fieldValues, value];
+                oldActiveFilters[field] = addFieldValue(fieldValues, value);
             } else {
                 oldActiveFilters[field] = removeValueFromValues(value, fieldValues)
-                /* if filter remains with no values remove it key. When checking filters if
-                 there is filters blank values keys will return that filters are truthy 
-                 (when they're nothing but just blank key) */
-                if (oldActiveFilters[field].length === 0) {
-                    delete oldActiveFilters[field]
-                }
+                removeFieldFromFiltersWhenValueAreEmpty(oldActiveFilters, field)
             }
             return { ...oldActiveFilters }
         })
 
-        handleFilter(activeFilters)
+        function removeFieldFromFiltersWhenValueAreEmpty(filters, field) {
+            if (filters[field].length === 0) {
+                delete filters[field]
+            }
+        }
+
+        function addFieldValue(values, value) {
+            return [...values, value]
+        }
 
         function removeValueFromValues(value, values) {
             return values.filter(v => v !== value)
@@ -115,7 +80,6 @@ export default function ProductsFilter(props) {
 
     function cleanFilters() {
         setActiveFilters([])
-        handleFilter(activeFilters)
     }
 
     return (
